requirements.txt
PyYAML>=6.0
jinja2>=3.1
requests>=2.31
beautifulsoup4>=4.12
scikit-learn>=1.4
joblib>=1.3

